// kaliversus_api/src/common/enums.ts
export enum UserRole {
  AUTOR = 'ROLE_AUTOR',
  REVISOR = 'ROLE_REVISOR',
  EDITOR = 'ROLE_EDITOR',
  ADMIN = 'ROLE_ADMIN',
  LECTOR = 'ROLE_LECTOR',
}

export enum EstadoPublicacion {
  BORRADOR = 'BORRADOR',
  EN_REVISION = 'EN_REVISION',
  CAMBIOS_SOLICITADOS = 'CAMBIOS_SOLICITADOS',
  APROBADO = 'APROBADO',
  PUBLICADO = 'PUBLICADO',
  RETIRADO = 'RETIRADO',
}

export enum TipoPublicacion {
  ARTICULO = 'ARTICULO',
  LIBRO = 'LIBRO',
}

export enum EstadoRevision {
  PENDIENTE = 'PENDIENTE',
  EN_PROCESO = 'EN_PROCESO',
  DEVUELTA = 'DEVUELTA',
  ACEPTADA = 'ACEPTADA',
  RECHAZADA = 'RECHAZADA',
  SUGERENCIA = 'SUGERENCIA',
}